Moved ErrorBoundary from containers folder to components folder
